% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorList.R
\name{addReadsToTree}
\alias{addReadsToTree}
\title{Add reads to the tree}
\usage{
addReadsToTree(
  fL,
  mappedFeaturesDF,
  featuresCol = "transcript_id",
  readsCol = NULL,
  unassigned = FALSE,
  extraTreeBranch = NULL,
  ...
)
}
\arguments{
\item{fL}{factorList of the annotation}

\item{mappedFeaturesDF}{\code{DFrame} with Sequences and transcript IDs}

\item{featuresCol}{Feature name column}

\item{readsCol}{Column with sequences, if omitted will use the row.names}

\item{unassigned}{If unassigned reads to be included in the tree or not}

\item{extraTreeBranch}{Additional sequences as \code{FactorList} to be added to
the tree. One can obtain unaligned reads using \code{getReadsFromBam} to obtain
reads from bam files.}

\item{...}{Other parameters to be passed to \code{fList2tree}}
}
\description{
Add reads to the tree
}
