% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{getOverlapValidity}
\alias{getOverlapValidity}
\title{getOverlapValidity}
\usage{
getOverlapValidity(sources, rules = defaultAssignRules())
}
\arguments{
\item{sources}{A data.frame of overlaps, as produced by
\code{\link{overlapWithTx}}}

\item{rules}{Assignment rules (see \code{\link{defaultAssignmentRules}})}
}
\value{
Data table (same format as \code{sources})
}
\description{
getOverlapValidity
}
