% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mylapply}
\alias{mylapply}
\title{Run \code{lapply} function in parallel, if possible.}
\usage{
mylapply(...)
}
\arguments{
\item{any}{parameters for \code{lapply} or \code{mclapply}}
}
\value{
A \code{list}.
}
\description{
Run \code{lapply} function in parallel, if possible.
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
